% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-CLX.R
\name{Cai.Mmatrix}
\alias{Cai.Mmatrix}
\title{The standardized statistics for (cov(Y)-cov(X))^2}
\usage{
Cai.Mmatrix(X, Y)
}
\arguments{
\item{X}{n1-by-p matrix, observation of the first population, columns are features}

\item{Y}{n2-by-p matrix, observation of the second population, columns are features}
}
\value{
A p-by-p numeric matrix, the estimation of (cov(Y)-cov(X))^2
}
\description{
A helper function to calculate the M-matrix in Cai et al. (2013), equation (2)
}
\seealso{
\code{Cai.max.test()}.
}

