% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-CLX.R
\name{Cai.theta.sigma}
\alias{Cai.theta.sigma}
\title{Estimation of cov(X) and its variance}
\usage{
Cai.theta.sigma(X)
}
\arguments{
\item{X}{n by p observation matrix, columns are features}
}
\value{
A list with the following components:
 \item{sigma}{A p-by-p matrix, estimation of cov(X)}
 \item{theta}{A p-by-p matrix, estimation of the variance of cov(X)}
}
\description{
A helper function to calculate an estimation of cov(X) (\code{sigma})
 and its variance (\code{theta}). See equation (2) in Cai (2013)
}
\seealso{
\code{Cai.max.test()}.
}

