% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-CZZW.R
\name{Chang.maxBoot.test}
\alias{Chang.maxBoot.test}
\title{Two-sample covariance test (Chang et al. 2016)}
\usage{
Chang.maxBoot.test(X, Y, nresample = 1000, useMC = TRUE, mc.cores = 1)
}
\arguments{
\item{X}{n1 by p matrix, observation of the first population, columns are features}

\item{Y}{n2 by p matrix, observation of the second population, columns are features}

\item{nresample}{the number of bootstraps to perform}

\item{useMC}{logical variable indicating whether to use multicore parallelization.
R packages \code{parallel} and \code{doParallel} are required if set to \code{TRUE}.}

\item{mc.cores}{decide the number of cores to use when \code{useMC} is set to \code{TRUE}.}
}
\value{
A list with the following components:
 \item{test.stat}{test statistic}
 \item{test.stat.boot}{bootstrap test statistics, a numeric vector with length nresample}
 \item{pVal}{bootstrap p-value}
}
\description{
The two-sample test for high-dimensional covariance matrices in
 Chang, Zhou, Zhou, and Wang (2016)
"Comparing Large Covariance Matrices under Weak Conditions on the Dependence Structure".
Using a wild boostrap procedure, the test statistic is essentially the square root of
 the statistic proposed in Cai, Liu and Xia (2013).
}
\references{
Chang, Zhou, Zhou, and Wang (2016)
"Comparing Large Covariance Matrices under Weak Conditions on the Dependence Structure",
arXiv preprint arXiv:1505.04493.
}
\seealso{
\code{Cai.max.test()}, \code{LC.U.test()}, \code{WL.randProj.test()}, \code{Schott.Frob.test()}.
}

