% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getDiffMatrix}
\alias{getDiffMatrix}
\title{A helper function to get the differential matrix
\deqn{D = A(Y) - A(X)}
where A() represents the weighted adjacency matrices defined as
\deqn{A_{ij} = |corr(i, j)|^b}
for some constant b > 0, 1 <= i, j <= p.
We let A represent the regular correlation matrix when b=0.}
\usage{
getDiffMatrix(X, Y, adj.beta)
}
\arguments{
\item{X}{n1-by-p matrix for samples from the first population. 
Rows are samples/observations, while columns are the features.}

\item{Y}{n2-by-p matrix for samples from the second population. 
Rows are samples/observations, while columns are the features.}

\item{adj.beta}{Power to transform correlation matrices to weighted adjacency matrice
by \eqn{A_{ij} = |r_ij|^adj.beta} where \eqn{r_ij} represents the Pearson correlation.
adj.beta=0 indicates using the regular correlation marix \eqn{(r_ij)}.}
}
\value{
Dmat the differential matrix \eqn{D = A(Y) - A(X)}
}
\description{
A helper function to get the differential matrix
\deqn{D = A(Y) - A(X)}
where A() represents the weighted adjacency matrices defined as
\deqn{A_{ij} = |corr(i, j)|^b}
for some constant b > 0, 1 <= i, j <= p.
We let A represent the regular correlation matrix when b=0.
}

