% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-permuteTest.R
\name{getPermutePval}
\alias{getPermutePval}
\title{Permutation p-value}
\usage{
getPermutePval(X, Y, testFunction, npermute = 100)
}
\arguments{
\item{X}{n1 by p matrix of the first population}

\item{Y}{n2 by p matrix of the second population}

\item{testFunction}{a function that takes in two arguments, X and Y,
and returns a list including the field \code{test.stat} for test statistic}

\item{npermute}{the number of permutations to conduct}
}
\value{
A list containing the following components:
 \item{test.res}{the results of the original test, including test statistic,
       sometimes also include the p-values calculated using the limiting distribution}
 \item{test.stat.permute}{a numeric vector with length \code{permute}, 
         for the test statsitics in permutations}
 \item{pVal}{the permutation pvalue, calculated as the fraction of \code{test.stat.permute}
     being larger than \code{test.res$test.stat}}
}
\description{
Obtain the p-value using permutation test.
}
\seealso{
\code{Cai.max.test()}, \code{Chang.maxBoot.test()}, 
       \code{LC.U.test()}, \code{WL.randProj.test()}, \code{Schott.Frob.test()}.
}

