% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-permuteTest.R
\name{getPermuteStat}
\alias{getPermuteStat}
\title{Permute test statistics}
\usage{
getPermuteStat(X, Y, testFunction, npermute = 100)
}
\arguments{
\item{X}{n1 by p matrix of the first population}

\item{Y}{n2 by p matrix of the second population}

\item{testFunction}{a function that takes in two arguments, X and Y,
and returns a list including the field \code{test.stat} for test statistic}

\item{npermute}{the number of permutations to conduct}
}
\value{
a numeric vector with length \code{npermute} for the test statsitics in permutations.
}
\description{
Obtain the null distribution of test statistics in permutation.
}

