% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-LC.R
\name{LC.Avalue}
\alias{LC.Avalue}
\title{U-statistic of \eqn{tr(cov(X)^2)}}
\usage{
LC.Avalue(X)
}
\arguments{
\item{X}{n by p matrix, columns are features, centered such that colMean=0}
}
\value{
A number, the U statistic for \eqn{tr(cov(X)^2)}.
}
\description{
Calculate the A value as in Li and Chen (2012) equation (2.1), 
for estimating \eqn{tr(cov(X)^2)}.
As suggested in the paper, only the first term is used after centering data.
}

