% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-LC.R
\name{LC.Cvalue}
\alias{LC.Cvalue}
\title{U-statistic of \eqn{tr(cov(X)cov(Y))}.}
\usage{
LC.Cvalue(X, Y)
}
\arguments{
\item{X}{n1 by p matrix of the first population, centered with colMean=0}

\item{Y}{n2 by p matrix of the second population, centered with colMean=0}
}
\value{
A number, the U statistic for \eqn{tr(cov(X)cov(Y))}.
}
\description{
Calculate the C value as in Li and Chen (2012) equation (2.2), 
for estimating \eqn{tr(cov(X)cov(Y))}.
As suggested in the paper, only the first term is used after centering data.
}

