% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-LC.R
\name{LC.U.test}
\alias{LC.U.test}
\title{Two-sample covariance test (Li and Chen 2012)}
\usage{
LC.U.test(X, Y)
}
\arguments{
\item{X}{n1 by p matrix, observation of the first population, columns are features}

\item{Y}{n2 by p matrix, observation of the second population, columns are features}
}
\value{
A list containing the following components:
 \item{Tn}{the U statistic for ||cov(X)-cov(Y)||_F^2}
 \item{Tn.sd}{the estimated standard deviation of \code{Tn}}
 \item{test.stat}{test statistic, \code{Tn}/\code{Tn.sd}}
 \item{pVal}{the p-value calculated using the limiting distribution (standard normal)}
}
\description{
Testing the equality of two high-dimensional covariance matrices
based on the \eqn{L_2} norm, proposed in Li and Chen (2012)
"Two Sample Tests for High-dimensional Covariance Matrices"
}
\seealso{
\code{Cai.max.test()}, \code{Chang.maxBoot.test()}, 
\code{WL.randProj.test()}, \code{Schott.Frob.test()}.
}

