% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-CZZW.R
\name{oneBootstrapStat}
\alias{oneBootstrapStat}
\title{One repetition in wild bootstrap}
\usage{
oneBootstrapStat(i, X.center, Y.center, n1, n2, sigma.x, sigma.y, Tdenominator)
}
\arguments{
\item{i}{i-th bootstrap}

\item{X.center}{n1 by p matrix, observation of the first population, 
columns are centered features}

\item{Y.center}{n2 by p matrix, observation of the second population, 
columns are centered features}

\item{n1}{number of rows of \code{X.center}}

\item{n2}{number of rows of \code{Y.center}}

\item{sigma.x}{p-by-p matrix, sample covariance of X, pre-calculated}

\item{sigma.y}{p-by-p matrix, sample covariance of X, pre-calculated}

\item{Tdenominator}{the denominator of the test statistics, pre-calculated}
}
\value{
The test statistic in this boostrap repetition.
}
\description{
Helper function for \code{Chang.wildBootstrap()}.
}
\seealso{
\code{Chang.maxBoot.test()}.
}

