% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-permuteTest.R
\name{onePermuteStat}
\alias{onePermuteStat}
\title{Test statistic in one permutation}
\usage{
onePermuteStat(i, Z, n1, n2, testFunction, seeds = NULL)
}
\arguments{
\item{i}{the i-th permutation, in the range of {1, ..., \code{npermute}}.}

\item{Z}{(n1+n2)-by-p matrix, containing the pooled samples from two populations. Columns are features.}

\item{n1}{the first n1 rows in Z represent the first population}

\item{n2}{the (n1+1):(n1+n2) rows in Z represent the second population}

\item{testFunction}{a function that takes in two arguments, X and Y,
and returns a list including the field \code{test.stat} for test statistic}

\item{seeds}{a numeric vector with length \code{npermute}, 
where \code{seeds[i]} specifies the seeding for the i-th permutation. 
Set to \code{NULL} if do not want to specify.}
}
\value{
Return the test statistic in the i-th permutation.
}
\description{
Helper function for \code{getPermutePval()} when using multi-core parallelization.
}
\seealso{
\code{getPermutePval()}.
}

