% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLEDpermute.R
\name{permuteIndex}
\alias{permuteIndex}
\title{Permute indices}
\usage{
permuteIndex(n1, n2)
}
\arguments{
\item{n1}{number of observations from first population}

\item{n2}{number of observations from second population}
}
\value{
A list with the following components:
 \item{i1}{a numeric vector with length \code{n1}, the indices for permuted first population}
 \item{i2}{a numeric vector with length \code{n2}, the indices for permuted second population}
}
\description{
Permute indices
}

