% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-Schott.R
\name{Schott.Frob.test}
\alias{Schott.Frob.test}
\title{Two-sample covariance test (Schott 2007)}
\usage{
Schott.Frob.test(X, Y)
}
\arguments{
\item{X}{n1 by p matrix, observation of the first population, columns are features}

\item{Y}{n2 by p matrix, observation of the second population, columns are features}
}
\value{
A list containing the following components:
 \item{test.stat}{test statistic}
 \item{pVal}{the p-value calculated using the limiting distribution (standrad normal)}
}
\description{
The two-sample covariance test proposed in Schott (2007) 
"A test for the equality of covariance matrices when
 the dimension is large relative to the sample sizes".
 The original test procedure handles multiple multivariate Normal distributions, 
 but for the purpose of our simulation, the implementation is for two populations only.
}
\seealso{
\code{Cai.max.test()}, \code{Chang.maxBoot.test()}, 
\code{LC.U.test()}, \code{WL.randProj.test()}.
}

