% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLEDpermute.R
\name{sLEDOnePermute}
\alias{sLEDOnePermute}
\title{One permutation of sLED}
\usage{
sLEDOnePermute(i, Z, n1, n2, seeds = NULL, sumabs.seq = 0.2, adj.beta = 0,
  rho = 1000, verbose = TRUE, niter = 20, trace = FALSE)
}
\arguments{
\item{i}{the i-th permutation}

\item{Z}{(n1+n2)-by-p matrix, containing the pooled samples from two populations. Columns are features.}

\item{n1}{the first n1 rows in Z represent the first population}

\item{n2}{the (n1+1):(n1+n2) rows in Z represent the second population}

\item{seeds}{a numeric vector with the length equals to npermute, specifying the
seed for randomly permutating samples in each permutation. 
If set tu NULL, then the default seeds are used.}

\item{sumabs.seq}{a numeric vector specifing the sequence of sparsity parameters to use, 
each between \eqn{1/sqrt(p)} and 1.}

\item{adj.beta}{a positive number representing the power to transform correlation matrices 
to weighted adjacency matrices by \eqn{A_{ij} = |r_ij|^adj.beta}, 
where \eqn{r_ij} represents the Pearson correlation.
When adj.beta=0, the regular correlation marix \eqn{(r_ij)} is used.}

\item{rho}{a large positive constant such that \eqn{A(X)-A(Y)+diag(rep(rho, p))} is positive definite.}

\item{verbose}{whether to print the progress during permutation tests}

\item{niter}{the number of iterations to use in the PMD algorithm (see \code{sLEDTestStat()})}

\item{trace}{whether to trace the progress of PMD algorithm (see \code{sLEDTestStat()})}
}
\value{
A list containing the following components:
 \item{Tn.permute}{A number represents the test statistic in this permutation.}
 \item{Tn.permute.sign}{A string, "pos" if the test statistic is given by sEig(D), and "neg" if is given by sEig(-D),
       where \code{sEig} denotes the sparse leading eigenvalue.}
}
\description{
One permutation of sLED
}

