% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLEDpermute.R
\name{sLEDpermute}
\alias{sLEDpermute}
\title{sLED permutation}
\usage{
sLEDpermute(Z, n1, n2, adj.beta = 0, rho = 1000, sumabs.seq = 0.2,
  npermute = 100, seeds = NULL, verbose = TRUE, niter = 20,
  trace = FALSE)
}
\arguments{
\item{Z}{(n1+n2)-by-p matrix, containing the samples from two populations with p features}

\item{n1}{the first n1 rows in Z represent the first population}

\item{n2}{the (n1+1):(n1+n2) rows in Z represent the second population}

\item{adj.beta}{a positive number representing the power to transform correlation matrices 
to weighted adjacency matrices by \eqn{A_{ij} = |r_ij|^adj.beta}, 
where \eqn{r_ij} represents the Pearson correlation.
When adj.beta=0, the regular covariance marix is used.}

\item{rho}{a large positive constant such that \eqn{A(X)-A(Y)+diag(rep(rho, p))} is positive definite.}

\item{sumabs.seq}{a numeric vector specifing the sequence of sparsity parameters to use, 
each between \eqn{1/sqrt(p)} and 1.}

\item{npermute}{number of permutations to use, default is 100}

\item{seeds}{a numeric vector with the length equals to npermute, specifying the
seed for randomly permutating samples in each permutation. 
If set tu NULL, then the default seeds are used.}

\item{verbose}{whether to print the progress during permutation tests}

\item{niter}{the number of iterations to use in the PMD algorithm (see \code{sLEDTestStat()})}

\item{trace}{whether to trace the progress of PMD algorithm (see \code{sLEDTestStat()})}
}
\value{
A list containing the following components:
 \item{Tn.perm}{a numeric vector with length \code{npermute}, the test statistic in permutations.}
 \item{Tn.perm.sign}{a vector of characters with length \code{npermute}, the sign in permutations: 
     "pos" if the permuted test statistic is given by sEig(D), and "neg" if is given by sEig(-D),
     where \code{sEig} denotes the sparse leading eigenvalue.}
}
\description{
Get the sLED test statistic on permuted samples.
}
\references{
Zhu, Lei, Devlin and Roeder (2016), "Testing High Dimensional Differential Matrices, 
with Application to Detecting Schizophrenia Risk Genes", arXiv:1606.00252.
}
\seealso{
\code{sLEDpermuteMC()}, \code{sLED()}.
}

