% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmPMD.R
\name{soft}
\alias{soft}
\title{Soft threshold}
\usage{
soft(x, d)
}
\arguments{
\item{x}{a numeric vector}

\item{d}{the soft threshold level}
}
\value{
the vector after soft thresholding \code{x} at level \code{d}.
}
\description{
Soft threshold
}
\seealso{
\code{symmPMD()}.
}

