% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-CZZW.R
\name{Chang.wildBootstrap}
\alias{Chang.wildBootstrap}
\title{Wild bootstrap}
\usage{
Chang.wildBootstrap(X, Y, sigma.x, sigma.y, Tdenominator, nresample = 1000)
}
\arguments{
\item{X}{n1 by p matrix, observation of the first population, columns are features}

\item{Y}{n2 by p matrix, observation of the second population, columns are features}

\item{Tdenominator}{the denominator of the test statistics, which is pre-calculated 
and remains unchanged in bootstrap.}

\item{nresample}{the number of bootstraps to perform}
}
\value{
A numeric vector with length \code{nresample}, containing the test statistics
in wild boostrap repetitions.
}
\description{
Perform the wild boostrap procedure for testing two-sample covariance
matrices in Chang et al. (2016).
}
\seealso{
\code{Chang.maxBoot.test()}.
}

