% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLED.R
\name{getDiffMatrix}
\alias{getDiffMatrix}
\title{The differential matrix}
\usage{
getDiffMatrix(X, Y, adj.beta = -1)
}
\arguments{
\item{X}{n1-by-p matrix for samples from the first population. 
Rows are samples/observations, while columns are the features.}

\item{Y}{n2-by-p matrix for samples from the second population. 
Rows are samples/observations, while columns are the features.}

\item{adj.beta}{Power to transform correlation matrices to weighted adjacency matrice
by \eqn{A_{ij} = |r_ij|^adj.beta} where \eqn{r_ij} represents the Pearson correlation.
When \code{adj.beta=0}, the correlation marix is used. 
When \code{adj.beta<0}, the covariance matrix is used.
The default value is \code{adj.beta=-1}.}
}
\value{
The p-by-p differential matrix \eqn{D = A(Y) - A(X)}
}
\description{
Given observations from two populations X and Y, 
compute the differential matrix
\deqn{D = A(Y) - A(X)}
where A() is the covariance matrix, or the weighted adjacency matrices defined as
\deqn{A_{ij} = |corr(i, j)|^b}
for some constant b > 0, 1 <= i, j <= p.
Let A represent the regular correlation matrix when b=0, and covariance matrix when b<0.
}

