% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-WL.R
\name{oneProj}
\alias{oneProj}
\title{F-statistic in one random projection}
\usage{
oneProj(i, Xsample, Ysample, n1, n2, p)
}
\arguments{
\item{i}{i-th repetition}

\item{Xsample}{n1 by p matrix, observation of the first population, columns are features}

\item{Ysample}{n2 by p matrix, observation of the second population, columns are features}
}
\description{
F-statistic in one random projection
}
\seealso{
\code{WL.randProj.test()}.
}

