% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLED.R
\name{sLED}
\alias{sLED}
\title{The sparse leading eigenvalue driven (sLED) test}
\usage{
sLED(X, Y, adj.beta = -1, rho = 1000, sumabs.seq = 0.2, npermute = 100,
  useMC = FALSE, mc.cores = 1, seeds = NULL, verbose = TRUE,
  niter = 20, trace = FALSE)
}
\arguments{
\item{X}{n1-by-p matrix for samples from the first population. 
Rows are samples/observations, while columns are the features.}

\item{Y}{n2-by-p matrix for samples from the second population. 
Rows are samples/observations, while columns are the features.}

\item{adj.beta}{a positive number representing the power to transform correlation matrices 
to weighted adjacency matrices by \eqn{A_{ij} = |r_ij|^adj.beta}, 
where \eqn{r_ij} represents the Pearson correlation.
When adj.beta=0, the correlation marix is used. 
When adj.beta<0, the covariance matrix is used.}

\item{rho}{a large positive constant such that \eqn{A(X)-A(Y)+diag(rep(rho, p))} is positive definite.}

\item{sumabs.seq}{a numeric vector specifing the sequence of sparsity parameters to use, 
each between \eqn{1/sqrt(p)} and 1.}

\item{npermute}{number of permutations to use, default is 100}

\item{useMC}{logical, whether to use multi-core version}

\item{mc.cores}{a number indicating how many cores to use in parallelization}

\item{seeds}{a numeric vector with the length equals to \code{npermute}, 
where \code{seeds[i]} specifies the seeding for the i-th permutation. 
Set to \code{NULL} if do not want to specify.}

\item{verbose}{whether to print the progress during permutation tests}

\item{niter}{the number of iterations to use in the PMD algorithm (see \code{symmPMD()})}

\item{trace}{logical, whether to trace the progress of PMD algorithm (see \code{symmPMD()})}
}
\value{
A list containing the following components:
 \item{Tn}{the test statistic}
 \item{Tn.perm}{the test statistic for permuted samples}
 \item{Tn.perm.sign}{the sign for permuted samples: "pos" if the permuted test statistic is given by sEig(D), 
         and "neg" if is given by sEig(-D)}
 \item{pVal}{the p-value of sLED test}  
 \item{sumabs.seq}{a numeric vector for a sequence of sparsity parameters. Default is 0.2.}
 \item{rho}{a positive constant to augment the diagonal of the differential matrix \eqn{D}
           such that \eqn{D + rho*I} becomes positive definite.}
 \item{stats}{a numeric vector of test statistics when using different sparsity parameters
           (corresponding to \code{sumabs.seq}).}
 \item{sign}{a vector of signs when using different sparsity parameters (corresponding to \code{sumabs.seq}).
         Sign is "pos" if the test statistic is given by sEig(D), and "neg" if is given by sEig(-D),
         where \code{sEig} denotes the sparse leading eigenvalue.}
 \item{v}{the sequence of sparse leading eigenvectors, each row corresponds to one sparsity 
         parameter given by \code{sumabs.seq}.}
 \item{leverage}{the leverage score for genes (defined as \eqn{v^2} element-wise) using 
         different sparsity parameters. Each row corresponds to one sparsity 
         parameter given by \code{sumabs.seq}.}
}
\description{
The sLED test for two-sample high-dimensional covariance and relationship matrices.
Suppose X, Y are p-dimensional random vectors independently coming from two populations.
Let \eqn{D} be the differential matrix given by
\deqn{D = A(Y) - A(X)} 
sLED tests the following hypothesis:
\deqn{H_0: D=0 versus H_1: D != 0}
where A() represents some p-by-p relationship matrix among features, including covariance matrices, 
correlation matrices, or the weighted adjacency matrices defined as
\deqn{A_{ij} = |corr(i, j)|^b}
for some constant b > 0, 1 <= i, j <= p. 
Let A represent the regular correlation matrix when b=0, and covariance matrix when b<0.
}
\references{
Zhu, Lei, Devlin and Roeder (2016), "Testing High Dimensional Differential Matrices, 
with Application to Detecting Schizophrenia Risk Genes", arXiv:1606.00252.
}
\seealso{
\code{symmPMD()}.
}

