% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sLEDTestStat.R
\name{sLEDTestStat}
\alias{sLEDTestStat}
\title{Calculate the sLED test statistic for differential matrices.}
\usage{
sLEDTestStat(Dmat, rho = 1000, sumabs.seq = 0.2, niter = 20,
  trace = FALSE)
}
\arguments{
\item{Dmat}{p-by-p numeric matrix, the differential matrix}

\item{rho}{a large positive constant such that \eqn{D+diag(rep(rho, p))} and \eqn{-D+diag(rep(rho, p))} 
are positive definite.}

\item{sumabs.seq}{a numeric vector specifing the sequence of sparsity parameters, each between \eqn{1/sqrt(p)} and 1.
Each sumabs*\eqn{$sqrt(p)$}{sqrt(p)} is the upperbound of the L_1 norm of leading sparse eigenvector \eqn{v}.}

\item{niter}{the number of iterations to use in the PMD algorithm (see \code{fastPMD()})}

\item{trace}{whether to trace the progress of PMD algorithm (see \code{fastPMD()})}
}
\value{
sumabs.seq  the sequence of sparsity parameters

rho the positive constant used

stats the sequence of test statistics

sign signs when using different sparsity parameters,
         "pos" if the test statistic is given by sEig(D), and "neg" if is given by sEig(-D).

v the sequence of sparse leading eigenvectors

leverage the leverage score for genes (defined as \eqn{v^2} element-wise) using different sparsity parameters
}
\description{
A differential matrix is the difference between two symmetric relationship matrices.
For any symmetric differential matrix \eqn{D}, sLED test statistic is defined as
\deqn{max{ sEig(D), sEig(-D) }}
where sEig() is the sparse leading eigenvalue, defined as
\deqn{$max_{v} v^T A v$}{max_{v} t(v)*A*v}
subject to
\deqn{$||v||_2 \leq 1, ||v||_1 \leq s$}{||v||_2 <= 1, ||v||_1 <= s}
}
\references{
Zhu, Lei, Devlin and Roeder (2016), "Testing High Dimensional Differential Matrices, 
with Application to Detecting Schizophrenia Risk Genes", arXiv:1606.00252.
}
\seealso{
\code{sLED()}, \code{fastPMD()}.
}

