% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-WL.R
\name{WL.randProj.test}
\alias{WL.randProj.test}
\title{Two-sample covariance test (Wu and Li 2015)}
\usage{
WL.randProj.test(X, Y, nproj = 100, useMC = FALSE, mc.cores = 1)
}
\arguments{
\item{X}{n1 by p matrix, observation of the first population, columns are features}

\item{Y}{n2 by p matrix, observation of the second population, columns are features}

\item{nproj}{number of random projections to use}

\item{useMC}{logical variable indicating whether to use multicore parallelization.
R packages \code{parallel} and \code{doParallel} are required if set to \code{TRUE}.}

\item{mc.cores}{decide the number of cores to use when \code{useMC} is set to \code{TRUE}.}
}
\value{
A list containing the following components:
 \item{test.stat}{test statistic}
 \item{pVal}{the p-value calculated using the limiting distribution 
           (max of independent standard normal)}
}
\description{
The two-sample covariance test using random projections 
proposed in Wu and Li (2015) 
"Tests for High-Dimensional Covariance Matrices Using Random Matrix Projection".
}
\references{
Wu and Li (2015) 
"Tests for High-Dimensional Covariance Matrices Using Random Matrix Projection",
arXiv preprint arXiv:1511.01611.
}
\seealso{
\code{Cai.max.test()}, \code{Chang.maxBoot.test()}, \code{LC.U.test()},
 \code{Schott.Frob.test()}.
}

